+++ src/index.php
@@ -140,7 +140,7 @@
             echo 'banner_key = "'.$config["banner_key"].'";';
         ?>
     </script>
-    <script id="bundle" src="https://cdn.chub.page/js/bundles/c841f3a9f7bbd1bb.js"></script>
+    <script id="bundle" src="https://cdn.chub.page/js/bundles/2dc6697020f6e4ba.js"></script>
 
     <?php
     $application_html = "";